Combine libs/ballot-interpretation test files
